+++ project/BuildSettings.scala
@@ -2,7 +2,7 @@ object BuildSettings {
 
   object version {
 
-    val airstream      = "0.12.0-M2"
+    val airstream      = "0.12.0-RC1"
     val `tuplez-apply` = "0.3.1"
     val utest          = "0.7.7"