+++ src/main/scala/io/frontroute/internal/Util.scala
@@ -1,0 +1,10 @@
+package io.frontroute
+package internal
+
+import com.raquo.airstream.core.EventStream
+
+private[frontroute] object Util {
+
+  private[frontroute] def rejected: EventStream[RouteResult] = EventStream.fromValue(RouteResult.Rejected)
+
+}