+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 
-### 0.12.0-M2
+### 0.12.0-RC1
 
-Update to Airstream `v0.12.0-M2`.
+Update to Airstream `v0.12.0-RC1`.
 
+* Breaking: the `io.frontroute.directives._` no longer exists and importing it is not needed anymore
 * API: renamed `RouteLocationProvider` into `LocationProvider`
 * API: new `CustomLocationProvider`
 * API: renamed `completeN` into `complete`
 * API: new `state` directive
+* API: new `makeRoute` and `makeRouteWithCallback`
+* Util: new `LocationUtils` object with `parseLocationParams` and `encodeLocationParams` functions
 * Bugfix: pop state event processing – event itself was used as `state`
+* Bugfix: link handler was not verifying if the link was pointing to the URL within the same origin, caused errors 
+  (changing then origin when pushing state is not allowed)
 
 ### 0.11.7